+++ multihop-lqf/general_network_simulator.h
@@ -17,8 +17,8 @@
 // files are queue lengths sampled at output periods.
 int general_network_simulator(
         const BoolMat &maximal_schedule_matrix,
-        int num_iterations,
-        int output_period,
+        int64_t num_iterations,
+        int64_t output_period,
         const Rates &arrival_rate,
         const std::string &filename_lqf,
         const std::string &filename_bp,
